Divide the processing of Gnavi API #15
